v1

+++ app/Http/Controllers/UserController.php
@@ -5,15 +5,14 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UserStoreRequest;
 use App\Http\Requests\UserUpdateRequest;
 use App\Http\Resources\UserResource;
-use App\Models\chef;
-use App\Models\role;
 use App\Models\User;
+use App\Traits\GeneralOutput;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
+    use GeneralOutput;
     function show(Request $request)
     {
         $filters = [];
@@ -23,42 +22,68 @@ class UserController extends Controller
         $request->filled('item_id') ? $filters[] = ['item_id', '=', $request->id] : 0;
         // $request->filled('role_id')?$filters[]=['role_id', '=', $request->id] : 0;
 
-        $ms = User::where($filters)
+        $user = User::where($filters)
             ->when($request->role_id != [], function ($q) use ($request) {
                 return $q->whereIn('role_id', $request->role_id);
             })->get();
 
-        return UserResource::collection($ms);
+        return $this->returnData(
+            'data',
+            UserResource::collection($user),
+        );
     }
 
     function store(UserStoreRequest $request)
     {
         $user = User::create($request->all());
         $request->filled('item_id') ? $user->chef()->create($request->only('item_id')) : 0;
-        return $user;
+        return $this->returnData(
+            'data',
+            $user,
+            'success create'
+        );
     }
     function update(UserUpdateRequest $request, $id)
     {
+        $validate = validator(['id' => $id], ['id' => 'exists:users,id']);
+        if ($validate->fails())
+            return $this->returnError($validate->errors()->getMessages());
+
         $user = User::findOrFail($id);
         $user->update($request->all());
         $request->filled('item_id') ? $user->chef()->update($request->only('item_id')) : 0;
         if (!$user->active) {
             $user->tokens()->delete();
         }
-        return $user;
+        return $this->returnData(
+            'data',
+            $user,
+            'success update'
+        );
     }
 
     function destroy($id)
     {
-        $item = User::findOrFail($id)->update(['active' => false]);
-        return $item;
+        $validate = validator(['id' => $id], ['id' => 'exists:users,id']);
+        if ($validate->fails())
+            return $this->returnError($validate->errors()->getMessages());
+
+        $user = User::findOrFail($id);
+        $user->update(['active' => false]);
+        $user->tokens()->delete();
+
+        return $this->returnData(
+            'data',
+            $user,
+            'success user status changed'
+        );
     }
 
     function login(Request $request)
     {
         $ability = ['0'];
-        $request->validate([
-            'email' => 'email|required',
+        $rr= $request->validate([
+            'email' => 'email|exists:users,email|required',
             'password' => 'string|required'
         ]);
 
@@ -71,23 +96,26 @@ class UserController extends Controller
                 $ability = ['editItem'];
                 break;
             case 3:
-                $ability = ['editItem','editOrder'];
+                $ability = ['editItem', 'editOrder'];
                 break;
             case 4:
                 $ability = ['editItem'];
                 break;
             default:
-               $ability=['showItem','showOrder'];
+                $ability = ['showItem', 'showOrder'];
         }
         if ($user->active && Hash::check($request->password, $user->password)) {
-            return $user->createToken('my', $ability)->plainTextToken;
+             $token=$user->createToken('my', $ability)->plainTextToken;
+             return $this->returnData('token',$token,'success login'); 
         } else {
-            return 'error';
+            return $this->returnError('error');
         }
     }
 
     function logout(Request $request)
     {
         $request->user()->currentAccessToken()->delete();
+        return $this->returnSuccessMessage("success logout");
+
     }
 }
